refactor(broker): extract abstract class for broker startup steps
