+++ broker/src/main/java/io/camunda/zeebe/broker/bootstrap/MonitoringServerStep.java
@@ -7,15 +7,11 @@
  */
 package io.camunda.zeebe.broker.bootstrap;
 
-import static io.camunda.zeebe.util.sched.future.CompletableActorFuture.completedExceptionally;
-
 import io.camunda.zeebe.broker.system.monitoring.BrokerHealthCheckService;
 import io.camunda.zeebe.util.sched.ConcurrencyControl;
 import io.camunda.zeebe.util.sched.future.ActorFuture;
-import io.camunda.zeebe.util.startup.StartupStep;
-import java.util.function.BiConsumer;
 
-final class MonitoringServerStep implements StartupStep<BrokerStartupContext> {
+final class MonitoringServerStep extends AbstractBrokerStartupStep {
 
   @Override
   public String getName() {
@@ -23,24 +19,7 @@ final class MonitoringServerStep implements StartupStep<BrokerStartupContext> {
   }
 
   @Override
-  public ActorFuture<BrokerStartupContext> startup(
-      final BrokerStartupContext brokerStartupContext) {
-    return createFutureAndRun(
-        brokerStartupContext,
-        (concurrencyControl, future) ->
-            startupInternal(brokerStartupContext, concurrencyControl, future));
-  }
-
-  @Override
-  public ActorFuture<BrokerStartupContext> shutdown(
-      final BrokerStartupContext brokerShutdownContext) {
-    return createFutureAndRun(
-        brokerShutdownContext,
-        (concurrencyControl, future) ->
-            shutdownInternal(brokerShutdownContext, concurrencyControl, future));
-  }
-
-  private void startupInternal(
+  void startupInternal(
       final BrokerStartupContext brokerStartupContext,
       final ConcurrencyControl concurrencyControl,
       final ActorFuture<BrokerStartupContext> startupFuture) {
@@ -56,22 +35,8 @@ final class MonitoringServerStep implements StartupStep<BrokerStartupContext> {
                 startupFuture));
   }
 
-  private void completeStartup(
-      final BrokerStartupContext brokerStartupContext,
-      final ActorFuture<BrokerStartupContext> startupFuture,
-      final BrokerHealthCheckService healthCheckService,
-      final Throwable error) {
-    if (error != null) {
-      startupFuture.completeExceptionally(error);
-    } else {
-      final var springBrokerBridge = brokerStartupContext.getSpringBrokerBridge();
-      springBrokerBridge.registerBrokerHealthCheckServiceSupplier(() -> healthCheckService);
-      brokerStartupContext.addPartitionListener(healthCheckService);
-      startupFuture.complete(brokerStartupContext);
-    }
-  }
-
-  private void shutdownInternal(
+  @Override
+  void shutdownInternal(
       final BrokerStartupContext brokerShutdownContext,
       final ConcurrencyControl concurrencyControl,
       final ActorFuture<BrokerStartupContext> shutdownFuture) {
@@ -94,32 +59,18 @@ final class MonitoringServerStep implements StartupStep<BrokerStartupContext> {
                 shutdownFuture));
   }
 
-  /**
-   * Helper function that tries to create a future and call a runnable. If an exception is thrown
-   * while creating the future, then this exception is forwarded to a dummy future
-   */
-  private ActorFuture<BrokerStartupContext> createFutureAndRun(
+  private void completeStartup(
       final BrokerStartupContext brokerStartupContext,
-      final BiConsumer<ConcurrencyControl, ActorFuture<BrokerStartupContext>> runnable) {
-    try {
-      final var concurrencyControl = brokerStartupContext.getConcurrencyControl();
-      final ActorFuture<BrokerStartupContext> future = concurrencyControl.createFuture();
-
-      forwardExceptions(() -> runnable.accept(concurrencyControl, future), future);
-      return future;
-    } catch (final Throwable t) {
-      return completedExceptionally(t);
-    }
-  }
-  /**
-   * helper function that forwards exceptions thrown by a synchronous block of code to a future
-   * object
-   */
-  private <V> void forwardExceptions(final Runnable r, final ActorFuture<V> future) {
-    try {
-      r.run();
-    } catch (final Throwable t) {
-      future.completeExceptionally(t);
+      final ActorFuture<BrokerStartupContext> startupFuture,
+      final BrokerHealthCheckService healthCheckService,
+      final Throwable error) {
+    if (error != null) {
+      startupFuture.completeExceptionally(error);
+    } else {
+      final var springBrokerBridge = brokerStartupContext.getSpringBrokerBridge();
+      springBrokerBridge.registerBrokerHealthCheckServiceSupplier(() -> healthCheckService);
+      brokerStartupContext.addPartitionListener(healthCheckService);
+      startupFuture.complete(brokerStartupContext);
     }
   }
 }
